refactor(team): parallel structure across headless providers (C23)
Two consistency-driven splits applying patterns proven in earlier PRs:

1. **Opencode MCP config split** — headless_opencode.go was 472
   lines bundling the turn body with MCP-config plumbing. The MCP
   helpers (writeHeadlessOpencodeMCPConfig, headlessOpencodeAgentConfigPath,
   safeHeadlessOpencodeConfigSlug, buildHeadlessOpencodeMCPEntry) move
   to a new headless_opencode_mcp.go (169 lines), mirroring the
   existing headless_openai_compat_mcp.go pattern. Now the two
   provider-specific MCP config flows are easy to compare side by
   side.

2. **Claude log helper consolidation** — appendHeadlessClaudeLog +
   appendHeadlessClaudeLatency move to headless_logging.go where C21
   already gathered the codex log helpers. headless_logging.go is
   now the single home for write-to-disk logging utilities used by
   every runner; previously the codex helpers lived there but the
   claude variants were stranded in headless_claude.go.

headless_opencode.go: 472 -> 319 lines (-153).
headless_claude.go: 317 -> 289 lines (-28).
New: headless_opencode_mcp.go (169 lines).
headless_logging.go: 74 -> 100 lines (+26 from claude appenders).

Stacked on refactor/launcher-c22-claudecommand-move (C22).
