+++ internal/team/headless_claude.go
@@ -5,9 +5,7 @@ import (
 	"context"
 	"fmt"
 	"io"
-	"os"
 	"os/exec"
-	"path/filepath"
 	"strings"
 	"time"
 
@@ -288,30 +286,4 @@ func (l *Launcher) buildHeadlessClaudeEnv(slug string) []string {
 		)
 	}
 	return env
-}
-
-func appendHeadlessClaudeLog(slug string, line string) {
-	dir := wuphfLogDir()
-	if dir == "" {
-		return
-	}
-	f, err := os.OpenFile(filepath.Join(dir, "headless-claude-"+slug+".log"), os.O_CREATE|os.O_WRONLY|os.O_APPEND, 0o600)
-	if err != nil {
-		return
-	}
-	defer func() { _ = f.Close() }()
-	_, _ = fmt.Fprintf(f, "[%s] %s\n", time.Now().Format(time.RFC3339), strings.TrimSpace(line))
-}
-
-func appendHeadlessClaudeLatency(slug string, line string) {
-	dir := wuphfLogDir()
-	if dir == "" {
-		return
-	}
-	f, err := os.OpenFile(filepath.Join(dir, "headless-claude-latency.log"), os.O_CREATE|os.O_WRONLY|os.O_APPEND, 0o600)
-	if err != nil {
-		return
-	}
-	defer func() { _ = f.Close() }()
-	_, _ = fmt.Fprintf(f, "[%s] agent=%s %s\n", time.Now().Format(time.RFC3339), strings.TrimSpace(slug), strings.TrimSpace(line))
 }
